requirements.txt
matplotlib
numpy
opencv_contrib_python
pandas
Pillow
pydicom
tqdm
utils
torch
torchvision
scikit-image
